Ensure postinstall script works on Windows
Hey, so I tried installing this on Windows and it didn’t work. I traced the error back to the postinstall script: it uses the OS's EOL, but in Windows this is `\r\n`, which is not ever going to be the line ending for Cypress’s `index.d.ts`; this in turn ended up not splitting any lines and the whole algorithm fails.

This PR changes all instances of `os.EOL` into `"\n"` which does the trick for both Windows and Linux. There should be no reason Cypress’s files are going to have OS-dependent line endings as far as I know, and Windows handles LF-endings just fine.

Thanks for the great lib!